WPF - Better Project Layout + MPV Control
