+++ Examples/WPF/Controls/MPVControl.xaml
@@ -1,0 +1,6 @@
+﻿<UserControl
+    x:Class="Nickvision.MPVSharp.Examples.WPF.Controls.MPVControl"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
+
+</UserControl>
