--- Examples/WPF/MPVControl.xaml
@@ -1,4 +1,0 @@
-﻿<UserControl
-    x:Class="Nickvision.MPVSharp.Examples.WPF.MPVControl"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml" />
