--- Examples/WPF/MainWindow.xaml
@@ -1,17 +1,0 @@
-﻿<Window x:Class="Nickvision.MPVSharp.Examples.WPF.MainWindow"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:Nickvision.MPVSharp.Examples.WPF"
-        mc:Ignorable="d"
-        Title="WPF MPV Example" Height="402" Width="640">
-    <Grid>
-        <Grid.RowDefinitions>
-            <RowDefinition Height="*"></RowDefinition>
-            <RowDefinition Height="Auto"></RowDefinition>
-        </Grid.RowDefinitions>
-        <local:MPVControl x:Name="MPVControl"/>
-        <Button Grid.Column="0" Grid.Row="1" Height="42" Click="OnPauseClicked">Toggle pause</Button>
-    </Grid>
-</Window>
