+++ Examples/WPF/Views/MainWindow.xaml
@@ -1,0 +1,20 @@
+﻿<Window x:Class="Nickvision.MPVSharp.Examples.WPF.Views.MainWindow"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:local="clr-namespace:Nickvision.MPVSharp.Examples.WPF.Views"
+		xmlns:controls="clr-namespace:Nickvision.MPVSharp.Examples.WPF.Controls"
+        mc:Ignorable="d"
+        Title="WPF MPV Example" Height="402" Width="640" Loaded="OnLoaded">
+    <Grid>
+        <Grid.RowDefinitions>
+            <RowDefinition Height="*"></RowDefinition>
+            <RowDefinition Height="42"></RowDefinition>
+        </Grid.RowDefinitions>
+        
+		<controls:MPVControl x:Name="MPV" Grid.Row="0"/>
+        
+		<Button Grid.Row="1" Height="42" Content="Toggle Pause" Click="OnPauseClicked"/>
+    </Grid>
+</Window>
